Setting some PEP8 standards in utils.py Fixing cycling dependencies due to methods being in other class
