+++ escaperoom/engine.py
@@ -2,6 +2,7 @@
 Stores the engine class and it's associated methods
 """
 from escaperoom.location import CurrentRoom, Item
+from escaperoom.transcript import Transcript
 
 
 class Engine:
@@ -186,7 +187,8 @@ class Engine:
         :return: Nothing
         """
         if self._inventory.is_inventory_complete():
-            with self._utils.open_file("final_gate.txt", "data") as final_gate_file:
+            with Transcript.open_file("final_gate.txt",
+                                      "data") as final_gate_file:
                 import re
                 group_id_pattern = re.compile(r"\s*group_id\s*=\s*([\w-]*)")
                 expected_hmac_pattern = re.compile(r"\s*expected_hmac\s*=\s*(\w*)")
